pyproject.toml
[tool.poetry]
name = "voxio"
version = "1.0.0"
description = "Library for ingesting and processing voxel (3D imaging) data"
authors = ["caniko <canhtart@gmail.com>"]
license = "BSD-4"
repository = "https://github.com/caniko/voxio"
readme = "README.md"

keywords = ["voxel", "processing", "fix", "mesh"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Visualization",
    "Operating System :: OS Independent"
]

[tool.poetry.scripts]
voxio = "voxio.cli:voxio_cli"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pydantic = "^2.0"
psutil = "^5.9.4"

# Maintained by project author
pydantic-numpy = ">=3.0"

numpy = ">=1.23"
scipy = "*"
scikit-image = "^0.20.0"
itk-morphologicalcontourinterpolation = "^1.1.0"
opencv-python = "^4.7.0.72"
imagesize = "^1.4.1"
click = "^8.1.3"
numba = "^0.56.4"
yaspin = "^2.3.0"
pillow = "^9.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.ruff]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
